POC: make torch_function handling in C++ more structured
This is one idea for preventing `should_skip_torch_function` from
being called multiple times inadvertantly. The idea is that you
implement torch function with three lambdas,
```
  return with_torch_function(
      /*checker=*/ [&](auto checker) {
        return checker.has_torch_function(self);
      },
      /*handle_torch_function=*/ [&] {
        return handle_torch_function(self, "_is_view", args);
      },
      /*default_impl=*/ [&] {
        auto& self_ = THPVariable_Unpack(self);
        return wrap(self_.is_view());
      });
```

By making `has_torch_function` the method of a class with a private
constructor, this means you can only access it from inside
`with_torch_function` and so the structure is forced on you.

ghstack-source-id: bb6e56792b2f9d74d7052f466613a1380ad02563
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83014