RFC: Don't introduce new overload for SymInt
Previously, we introduced new SymInt overloads for every function we
wanted.  This to a lot of boilerplate, and also a lot of confusion
about how the overloads needed to be implemented.

This PR takes a simpler but more risky approach: just take the original
function and changes its ints to SymInts.

This is BC-breaking in the following ways:

* The C++ API for registering implementations for aten operators will
  change from int64_t to SymInt whenever you make this change.
  Code generated registrations in PyTorch do not change as codegen
  handles the translation automatically, but manual registrations will
  need to follow the change.  Typically, if you now accept a SymInt
  where you previously only took int64_t, you have to convert it back
  manually.  This will definitely break XLA.

This is not BC-breaking in the following ways:

* The user facing C++ API remains compatible.  Even if a function
  changes from int to SymInt, the default C++ binding still takes only
  ints.  (e.g., at::empty(IntArrayRef, ...).  To call with SymInts, you
  must call at::empty_symint instead.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: 2bd263e9a28f4bffcd26aa47390277ba60354153
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83628